system OS differentiation and updates on Featuresfiles

+++ src/main/java/factory/DriverFactory.java
@@ -12,8 +12,9 @@ public class DriverFactory {
 	
 	public static ThreadLocal<WebDriver> threadDriver = new ThreadLocal<>();
 	
-	public WebDriver init_driver(String browser, String profile) throws IOException {
-		if(browser.equalsIgnoreCase("Chrome")) {
+	public WebDriver init_driver(String browser, String profile, String os) throws IOException {
+
+		if(browser.equalsIgnoreCase("Chrome") && os.contains("Windows")) {
 			Runtime rt = Runtime.getRuntime();
 			Process proc = rt.exec("taskkill /im chrome.exe /f /t");
 //			WebDriverManager.chromedriver().setup();
@@ -26,10 +27,30 @@ public class DriverFactory {
 			ChromeDriver driver = new ChromeDriver(options);
 			threadDriver.set(driver);
 		}
+
+		else if(browser.equalsIgnoreCase("Chrome") && os.contains("Mac")){
+
+			Runtime rt = Runtime.getRuntime();
+			Process proc = rt.exec("taskkill /im chrome.exe /f /t");
+//			WebDriverManager.chromedriver().setup();
+			System.setProperty("webdriver.chrome.driver", "./Drivers/chromedriver-win64/chromedriver.exe");
+			ChromeOptions options = new ChromeOptions();
+//			options.addArguments("--user-data-dir="+profile);
+			//options.addArguments("--profile-directory=Person 1");
+			options.addArguments("--start-maximized");
+
+			ChromeDriver driver = new ChromeDriver(options);
+			threadDriver.set(driver);
+
+		}
+
+
 		getDriver().manage().deleteAllCookies();
 		getDriver().manage().window().maximize();
 		return getDriver();
 	}
+
+
 	
 	public static synchronized WebDriver getDriver() {
 		return threadDriver.get();
